Sharing common modules across multiple remote applications
