+++ products/webpack.config.js
@@ -16,6 +16,7 @@ module.exports = {
       exposes: {
         "./ProductsIndex": "./src/index.js",
       },
+      shared: ['faker']
     }),
   ],
 };
